+++ common/src/main/java/me/drex/antixray/common/util/Arguments.java
@@ -1,6 +1,7 @@
 package me.drex.antixray.common.util;
 
 import me.drex.antixray.common.interfaces.IClientboundChunkBatchStartPacket;
+import net.minecraft.server.network.ServerGamePacketListenerImpl;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.chunk.ChunkAccess;
 
@@ -14,4 +15,5 @@ public class Arguments {
     public static final ThreadLocal<Integer> CHUNK_SECTION_INDEX = new ThreadLocal<>();
     public static final ThreadLocal<Object[]> PRESET_VALUES = new ThreadLocal<>();
     public static final ThreadLocal<IClientboundChunkBatchStartPacket> BATCH_START_PACKET = new ThreadLocal<>();
+    public static final ThreadLocal<ServerGamePacketListenerImpl> PACKET_LISTENER = new ThreadLocal<>();
 }