+++ common/src/main/java/me/drex/antixray/common/util/Util.java
@@ -3,6 +3,8 @@ package me.drex.antixray.common.util;
 import me.drex.antixray.common.interfaces.ILevel;
 import me.drex.antixray.common.interfaces.IPacket;
 import me.drex.antixray.common.util.controller.ChunkPacketBlockController;
+import me.drex.antixray.common.util.controller.DisabledChunkPacketBlockController;
+import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.network.protocol.Packet;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.LevelHeightAccessor;
@@ -13,6 +15,11 @@ public final class Util {
 
     public static ChunkPacketBlockController getBlockController(Level level) {
         return ((ILevel) level).getChunkPacketBlockController();
+    }
+
+    public static ChunkPacketBlockController getBlockController(ServerPlayer player) {
+        ChunkPacketBlockController controller = getBlockController(player.level());
+        return controller.shouldModify(player) ? controller : DisabledChunkPacketBlockController.NO_OPERATION_INSTANCE;
     }
 
     // Converts height accessors from ChunkAccess into levels.