+++ common/src/main/java/me/drex/antixray/common/util/controller/ChunkPacketBlockControllerAntiXray.java
@@ -13,10 +13,7 @@ import net.minecraft.core.Direction;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.network.protocol.game.ClientboundLevelChunkWithLightPacket;
 import net.minecraft.network.protocol.game.ServerboundPlayerActionPacket;
-import net.minecraft.server.level.ChunkHolder;
-import net.minecraft.server.level.ServerChunkCache;
-import net.minecraft.server.level.ServerLevel;
-import net.minecraft.server.level.ServerPlayerGameMode;
+import net.minecraft.server.level.*;
 import net.minecraft.world.level.ChunkPos;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.biome.Biomes;
@@ -43,6 +40,7 @@ public abstract class ChunkPacketBlockControllerAntiXray implements ChunkPacketB
     protected final int maxBlockHeight;
     private final int maxBlockHeightUpdatePosition;
     private final int updateRadius;
+    private final boolean usePermission;
     private final Object2BooleanOpenHashMap<BlockState> solidGlobal = new Object2BooleanOpenHashMap<>(Block.BLOCK_STATE_REGISTRY.size());
     private final Object2BooleanOpenHashMap<BlockState> obfuscateGlobal = new Object2BooleanOpenHashMap<>(Block.BLOCK_STATE_REGISTRY.size());
     private final LevelChunkSection[] emptyNearbyChunkSections = {EMPTY_SECTION, EMPTY_SECTION, EMPTY_SECTION, EMPTY_SECTION};
@@ -50,9 +48,10 @@ public abstract class ChunkPacketBlockControllerAntiXray implements ChunkPacketB
     // If an ExecutorService with multiple threads is used, ThreadLocal must be used here
     private final ThreadLocal<int[]> presetBlockStateBits = ThreadLocal.withInitial(() -> new int[getPresetBlockStatesLength()]);
 
-    protected ChunkPacketBlockControllerAntiXray(Level level, Set<Block> toObfuscate, int maxBlockHeight, int updateRadius, boolean lavaObscures) {
+    protected ChunkPacketBlockControllerAntiXray(Level level, Set<Block> toObfuscate, int maxBlockHeight, int updateRadius, boolean lavaObscures, boolean usePermission) {
         this.maxBlockHeight = maxBlockHeight;
         this.updateRadius = updateRadius;
+        this.usePermission = usePermission;
 
         for (Block block : toObfuscate) {
             // Don't obfuscate air because air causes unnecessary block updates and causes block updates to fail in the void
@@ -73,6 +72,11 @@ public abstract class ChunkPacketBlockControllerAntiXray implements ChunkPacketB
 
         this.maxBlockHeightUpdatePosition = maxBlockHeight + updateRadius - 1;
 
+    }
+
+    @Override
+    public boolean shouldModify(ServerPlayer player) {
+        return !this.usePermission || (!player.hasPermissions(2) && !AntiXray.INSTANCE.hasBypassPermission(player));
     }
 
     @Override