+++ common/src/main/java/me/drex/antixray/common/util/controller/ObfuscateLayerChunkPacketBlockController.java
@@ -8,8 +8,8 @@ import java.util.function.IntSupplier;
 
 public class ObfuscateLayerChunkPacketBlockController extends ObfuscateChunkPacketBlockController {
 
-    public ObfuscateLayerChunkPacketBlockController(Level level, Set<Block> replacementBlocks, Set<Block> hiddenBlocks, int maxBlockHeight, int updateRadius, boolean lavaObscures) {
-        super(level, replacementBlocks, hiddenBlocks, maxBlockHeight, updateRadius, lavaObscures);
+    public ObfuscateLayerChunkPacketBlockController(Level level, Set<Block> replacementBlocks, Set<Block> hiddenBlocks, int maxBlockHeight, int updateRadius, boolean lavaObscures, boolean usePermission) {
+        super(level, replacementBlocks, hiddenBlocks, maxBlockHeight, updateRadius, lavaObscures, usePermission);
     }
 
     @Override