+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Bypass permission
+
 ## [1.4.6] - 2024-08-16
 ### Fixed
 - Wrong argument